requirements.txt
numpy==1.13.1
Pillow==4.3.0
scikit-learn==0.19.0
scipy==0.19.1
tensorflow==1.3.0
